Include lopf and pf_post_lopf in execute.py
